[providers] - refactor: centralize provider configurations
 - Extract provider configurations for models and services into a central location for better management and reusability
 - Remove duplicate provider configuration code from the providers page and reference the centralized configurations instead
